Structure tests to run multi scenarios.
