+++ hc-dna/zomes/tests/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "RUST_LOG=debug TRYORAMA_LOG_LEVEL=debug TRYORAMA_HOLOCHAIN_PATH=\"holochain\" TIMEOUT=40000 TRYORAMA_ZOME_CALL_TIMEOUT_MS=100000 TRYORAMA_CONDUCTOR_TIMEOUT_MS=100000 ts-node index.ts",
+    "test": "RUST_LOG=debug TRYORAMA_LOG_LEVEL=debug TRYORAMA_HOLOCHAIN_PATH=\"holochain\" TIMEOUT=40000 TRYORAMA_ZOME_CALL_TIMEOUT_MS=100000 TRYORAMA_CONDUCTOR_TIMEOUT_MS=100000 ts-node index.ts | tap-diff",
     "test-signals": "RUST_LOG=debug TRYORAMA_LOG_LEVEL=debug TRYORAMA_HOLOCHAIN_PATH=\"holochain\" TIMEOUT=40000 TRYORAMA_ZOME_CALL_TIMEOUT_MS=100000 TRYORAMA_CONDUCTOR_TIMEOUT_MS=100000 ts-node signal-test.ts",
     "test-basic": "RUST_LOG=debug TRYORAMA_LOG_LEVEL=debug TRYORAMA_HOLOCHAIN_PATH=\"holochain\" TIMEOUT=40000 TRYORAMA_ZOME_CALL_TIMEOUT_MS=100000 TRYORAMA_CONDUCTOR_TIMEOUT_MS=100000 ts-node basic-test.ts",
     "test-pagination": "RUST_LOG=debug,holochain_p2p::spawn::actor=error,holochain::core::workflow::integrate_dht_ops_workflow=error,holochain::core::workflow::publish_dht_ops_workflow=error,wasmer_compiler_cranelift::translator::func_translator=error,hc_time_index::dfs::methods=error TRYORAMA_LOG_LEVEL=debug TRYORAMA_HOLOCHAIN_PATH=\"holochain\" TIMEOUT=40000 TRYORAMA_ZOME_CALL_TIMEOUT_MS=100000 TRYORAMA_CONDUCTOR_TIMEOUT_MS=100000 ts-node pagination-test.ts",
@@ -22,6 +22,7 @@
     "blake2b": "^2.1.3",
     "lodash": "^4.17.19",
     "tape": "^5.0.1",
+    "tap-diff": "^0.1.1",
     "ts-node": "^8.10.2",
     "typescript": "^4.2.4"
   }
